+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## Development
 ### 2018-10-26
 * Blocks, Threads: changed SPLIT "csv" to comply with RFC 4180, took out others options
+* fixed #2235 (help for local custom blocks)
 
 ### 2018-10-24
 * Blocks, Threads: added "csv records" to SPLIT options, renamed "csv" to "csv fields"